+++ src/index.js
@@ -1,1 +1,2 @@
 export default from './SearchApi'
+export { INDEX_MODES } from './util'