+++ src/util/index.js
@@ -1,1 +1,2 @@
 export default from './SearchUtility'
+export { INDEX_MODES } from './constants'